# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RedSecTools API",
    "version": "2.0.0",
    "description": "Repository-only OpenAPI documentation for RedSecTools. This file is not served by the live application. It documents the current HTTP API, including auth requirements for main app, guest-link, extension, and admin routes. RedSecThreat now includes user-scoped keyword, tag, and alert ownership on top of shared feed ingestion, plus admin-managed feed, template, and notification policy routes. Survey and Wiki include their active create/manage/respond/results/documentation route groups alongside the existing Paste, Share, Vault, Chat, Calendar, Homepage, Extension, and Admin surfaces."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_session",
        "description": "Signed user session cookie for the main web app."
      },
      "GuestSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_guest",
        "description": "Signed guest cookie created from /guest/{token}. Tool-scoped for paste or share."
      },
      "ExtensionBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque token",
        "description": "Bearer token returned by /api/ext/auth/login or /api/ext/auth/login/mfa."
      },
      "AdminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_admin",
        "description": "Signed admin session cookie for /admin/api routes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "retryAfter": { "type": "integer" },
          "restartLogin": { "type": "boolean" }
        },
        "required": ["error"]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        },
        "required": ["success"]
      },
...
      "SurveySummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "ownerId": { "type": "string" },
          "responseMode": { "type": "string" },
          "status": { "type": "string" },
          "publicToken": { "type": ["string", "null"] },
          "startsAt": { "type": ["integer", "null"] },
          "endsAt": { "type": ["integer", "null"] },
          "createdAt": { "type": "integer" },
          "updatedAt": { "type": "integer" }
        },
        "required": ["id", "title", "ownerId", "responseMode", "status"]
      },
...
    },
    "responses": {
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Success" }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or expired",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/survey/list": {
      "get": {
        "tags": ["Survey"],
        "summary": "List surveys visible to the current user",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Survey list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "surveys": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SurveySummary" }
                    }
                  },
                  "required": ["surveys"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/survey": {
      "post": {
        "tags": ["Survey"],
        "summary": "Create a survey",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/survey/{id}": {
      "get": {
        "tags": ["Survey"],
        "summary": "Get a survey and its questions",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey detail",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "put": {
        "tags": ["Survey"],
        "summary": "Update a survey",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "delete": {
        "tags": ["Survey"],
        "summary": "Delete a survey",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/{id}/status": {
      "put": {
        "tags": ["Survey"],
        "summary": "Change survey lifecycle state",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": ["publish", "close", "end_early", "reopen"]
                  }
                },
                "required": ["action"]
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/{id}/questions/reorder": {
      "put": {
        "tags": ["Survey"],
        "summary": "Reorder survey questions",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                },
                "required": ["order"]
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/{id}/stats": {
      "get": {
        "tags": ["Survey"],
        "summary": "Get survey statistics",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey statistics",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/{id}/results": {
      "get": {
        "tags": ["Survey"],
        "summary": "Get survey results",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey results",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/{id}/results/export": {
      "get": {
        "tags": ["Survey"],
        "summary": "Export survey results as CSV",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV export",
            "content": {
              "text/csv": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/{id}/responses/{responseId}": {
      "get": {
        "tags": ["Survey"],
        "summary": "Get a single survey response",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "responseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey response detail",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/survey/respond/{token}": {
      "get": {
        "tags": ["Survey"],
        "summary": "Get public survey response form data",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey response metadata",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/BadRequest" },
          "410": {
            "description": "Survey is closed or not yet open",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Survey"],
        "summary": "Submit a public or internal survey response",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/BadRequest" },
          "410": {
            "description": "Survey is closed or not yet open",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}